TS-19.code-ui-refactor. Fix waybill table
