+++ app/views/waybills/index.html.erb
@@ -1,1 +1,5 @@
-<%=react_component('Waybill',{}) %> 
+<center>
+  <div class="client_content">
+    <%=react_component('Waybill',{}) %> 
+  </div>
+</center>
